image upload now working between my fend and bend :
