+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 development_documentation/
 logs/
 eng.traineddata
+uploads/
